typedef unsigned long long

+++ headers/Stack/Stack.h
@@ -4,7 +4,7 @@
 #ifndef STACK_H_
 #define STACK_H_
 
-typedef double elem_t;
+typedef unsigned long long elem_t;
 
 const ssize_t ReallocationScale = 2;
 const ssize_t InitialCapacity = 10;
